+++ AccountsReceivable.API/Models/CustomerWalletTransaction.cs
@@ -11,7 +11,7 @@ namespace AccountsReceivable.API.Models
         [Key]
         public int CustomerWalletTransactionId { get; set; }
         public int? CustomerWalletId { get; set; }
-        public int? TransactionAmount { get; set; }
+        public decimal? TransactionAmount { get; set; }
         public string TransactionType { get; set; }
         public int? TransactionModeId { get; set; }
         public DateTime TransactionDate { get; set; }
